+++ packages/cloud-cognitive/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.22.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-products@1.21.0...@carbon/ibm-products@1.22.0) (2022-07-05)
+
+
+### Features
+
+* added center aligned columns component and centered new columns ([#2056](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2056)) ([9178df2](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/9178df24c2f54370ab4064ab659c77a9ac459377))
+
+
+
+
+
+# [1.21.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-products@1.20.3...@carbon/ibm-products@1.21.0) (2022-06-28)
+
+
+### Features
+
+* **WebTerminal:** Changes width to 640px ([#2066](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2066)) ([b741316](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/b741316b74dd8be0af2f4cfc61ad28c0a672e46b))
+
+
+
+
+
 ## [1.20.3](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-products@1.20.2...@carbon/ibm-products@1.20.3) (2022-06-21)
 
 **Note:** Version bump only for package @carbon/ibm-products